requirements.txt
numpy>=1.21.0
soundfile>=0.10.3
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
PyQt6>=6.4.0
pyqtgraph>=0.13.3
sounddevice>=0.4.5
pynput>=1.7.6
pyopengl>=3.1.6
pyqtgraph>=0.13.3 